CI/CD + 两阶段流水线 + 编译修复 + 阻止自动更新
- 两阶段流水线：check 全过 → release 构建
- clippy 合并到 Ubuntu check
- 预生成 proto + meval 替换 kalk
- telemetry 空壳完善（shell/pager 对齐）
- 阻止自动更新
- README 中英双版

+++ crates/codegen/xai-grok-telemetry/src/session_ctx.rs
@@ -171,7 +171,7 @@ fn write_event<T: Serialize>(
 /// never uploads.
 pub fn log_event<T: TelemetryEvent>(data: T) {
     let ctx = TELEMETRY_CTX.try_with(|c| {
-        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g as usize).ok())
+        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g).ok())
     }).ok();
     write_event(T::NAME, EmitterOrigin::Shell.event_prefix(), &data, ctx);
 }
@@ -184,7 +184,7 @@ pub fn log_event_dual<T: TelemetryEvent>(_internal_enabled: bool, data: T) {
 /// Session lifecycle event → local log.
 pub fn log_session_event<T: TelemetryEvent>(data: T) {
     let ctx = TELEMETRY_CTX.try_with(|c| {
-        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g as usize).ok())
+        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g).ok())
     }).ok();
     write_event(T::NAME, EmitterOrigin::Shell.event_prefix(), &data, ctx);
 }
@@ -192,7 +192,7 @@ pub fn log_session_event<T: TelemetryEvent>(data: T) {
 /// Session lifecycle event with origin → local log.
 pub fn log_session_event_with_origin<T: TelemetryEvent>(origin: EmitterOrigin, data: T) {
     let ctx = TELEMETRY_CTX.try_with(|c| {
-        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g as usize).ok())
+        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g).ok())
     }).ok();
     write_event(T::NAME, origin.event_prefix(), &data, ctx);
 }
@@ -201,7 +201,7 @@ pub fn log_session_event_with_origin<T: TelemetryEvent>(origin: EmitterOrigin, d
 pub fn emit_event<T: Serialize + Send + 'static>(event_suffix: impl Into<String>, data: T) {
     let event_name = format!("{}{}", EmitterOrigin::Shell.event_prefix(), event_suffix.into());
     let ctx = TELEMETRY_CTX.try_with(|c| {
-        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g as usize).ok())
+        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g).ok())
     }).ok();
     write_event(&event_name, EmitterOrigin::Shell.event_prefix(), &data, ctx);
 }
@@ -214,7 +214,7 @@ pub fn emit_event_with_origin<T: Serialize + Send + 'static>(
 ) {
     let event_name = format!("{}{}", origin.event_prefix(), event_suffix.into());
     let ctx = TELEMETRY_CTX.try_with(|c| {
-        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g as usize).ok())
+        (c.session_id.clone(), c.prompt_index.try_lock().map(|g| *g).ok())
     }).ok();
     write_event(&event_name, origin.event_prefix(), &data, ctx);
 }
